Add functions check if coordinates is occupied
the functions will set data-info to free on dragged item coordinates
list on drag start. the item will not be dropped if the drop coordinates
is occupied. the item will successfully drop on free coordinates and
change the previous coordinates data-info to free while the new
coordinates will be set to occupied"